requirements.txt
Django==1.10.2

# channels 
channels
asgi-redis
daphne

# REST
djangorestframework==3.4.6
djangorestframework-filters==0.8.0
djangorestframework-jwt==1.8.0
djoser==0.5.1
django-extra-fields

Pillow==3.2.0

# dev
ipdb

uWSGI==2.0.12
dj-database-url==0.4.1
psycopg2==2.6.1

whitenoise
pytest-django==2.9.1

graphene-django>=1.0
shortuuid==0.4.2
django-filer==1.2.5
django-image-cropping==1.0.3
